Types / code cleanup
