+++ src/constants/enums/contactRoles.ts
@@ -1,0 +1,13 @@
+const contactRoles = [
+  'Product Owner',
+  'System Owner',
+  'System Maintainer',
+  "Contracting Officer's Representative (COR)",
+  'Cloud Navigator',
+  'Privacy Advisor',
+  'CRA',
+  'Other',
+  'Unknown'
+];
+
+export default contactRoles;
